pyproject.toml
[project]
name = "BoldSystems-agent"
version = "0.1.0"
description = "Bold Systems resolver - DNA barcode related "
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Rohit Bogulla" }
]

dependencies = [
    # Core agent SDK
    "ichatbio-sdk>=0.2.5",

    # API server
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.41.0",

    # OpenAI + structured output
    "openai>=2.24.0",
    "instructor>=1.14.5",

    # Scientific stack

    # HTTP / utilities
    "httpx>=0.28.0",
    "requests>=2.32.0",
    "PyYAML>=6.0",
    "rich>=14.0.0",
    "typer>=0.24.0",

    # Config
    "python-dotenv>=1.0.1",

    # Validation
    "pydantic>=2.12.0",

    "rapidfuzz>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["setuptools>=80.0.0", "wheel"]
build-backend = "setuptools.build_meta"
